third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/scripts-panel.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script src="../../../http/tests/inspector/workspace-test.js"></script>
<script>
function test()
{
    function dumpNavigator(sourcesNavigatorView)
    {
        InspectorTest.addResult("Navigator:");
        InspectorTest.dumpNavigatorView(sourcesNavigatorView);
    }

    function createNavigatorView()
    {
        var workspace = Workspace.workspace;
        Workspace.workspace = InspectorTest.testWorkspace;
        var navigatorView = new Sources.SourcesNavigatorView();
        Workspace.workspace = workspace;
        navigatorView.show(UI.inspectorView.element);
        return navigatorView;
    }

    function createContentProvider(url)
    {
        var contentProvider = Common.StaticContentProvider.fromString(url, Common.resourceTypes.Script, "");
        contentProvider.requestContent = function()
        {
            InspectorTest.addResult("Source requested for " + url);
            return Promise.resolve("");
        };
        return contentProvider;
    }

    function createMockWorkspace()
    {
        InspectorTest.createWorkspaceWithTarget(true);
        InspectorTest.testDebuggerProject = new Bindings.ContentProviderBasedProject(InspectorTest.testWorkspace, "", Workspace.projectTypes.Debugger, "");
        return InspectorTest.testWorkspace;
    }

    function addNetworkFile(workspace, url)
    {
        return InspectorTest.testNetworkProject.addFile(createContentProvider(url), InspectorTest.mainFrame());
    }

    function addDebuggerFile(workspace, url)
    {
        var uiSourceCode = InspectorTest.testDebuggerProject.createUISourceCode(url, Common.resourceTypes.Script);
        InspectorTest.testDebuggerProject.addUISourceCodeWithProvider(uiSourceCode, createContentProvider(url));
        return uiSourceCode;
    }

    InspectorTest.runTestSuite([
        function testInitialLoad(next)
        {
            var workspace = createMockWorkspace();

            addNetworkFile(workspace, "foobar.js");

            var sourcesNavigatorView = createNavigatorView();

            addNetworkFile(workspace, "foo.js");
            addNetworkFile(workspace, "bar.js");
            var uiSourceCode = addNetworkFile(workspace, "baz.js");
            sourcesNavigatorView.revealUISourceCode(uiSourceCode);

            dumpNavigator(sourcesNavigatorView);
            next();
        },

        function testReset(next)
        {
            var workspace = createMockWorkspace();
            var sourcesNavigatorView = createNavigatorView();

            var uiSourceCode = addNetworkFile(workspace, "foo.js");
            addNetworkFile(workspace, "bar.js");
            addNetworkFile(workspace, "baz.js");

            dumpNavigator(sourcesNavigatorView);
            InspectorTest.addResult("Revealing in navigator.");
            sourcesNavigatorView.revealUISourceCode(uiSourceCode);
            dumpNavigator(sourcesNavigatorView);

            InspectorTest.testNetworkProject._reset();
            dumpNavigator(sourcesNavigatorView);
            uiSourceCode = addNetworkFile(workspace, "bar.js");
            sourcesNavigatorView.revealUISourceCode(uiSourceCode);
            dumpNavigator(sourcesNavigatorView);

            next();
        },

        function testDebuggerUISourceCodeAddedAndRemoved(next)
        {
            var workspace = createMockWorkspace();
            var sourcesNavigatorView = createNavigatorView();

            var uiSourceCode = addNetworkFile(workspace, "foo.js");
            var debuggerUISourceCode = addDebuggerFile(workspace, "compiled.js");
            sourcesNavigatorView.revealUISourceCode(uiSourceCode);
            sourcesNavigatorView.revealUISourceCode(debuggerUISourceCode);
            dumpNavigator(sourcesNavigatorView);

            // Plug compiler source mapping.
            addNetworkFile(workspace, "source.js");

            dumpNavigator(sourcesNavigatorView);
            next();
        }
    ]);
};
</script>
</head>
<body onload="runTest()">
<p>Tests that scripts panel UI elements work as intended.</p>
</body>
</html>
